FB: FIMD: set the alpha value width
This patch adds function which can set the alpha value width according
to transparency value of each window format.

The default value of transparency is changed from 0 to 8 in order to
set the alpha value width as 8. This is because it can be helpful to
prevent gradation. Also, FIMD hardware can support 8 bit alpha value.

This value can be changed when using FBIOPUT_VSCREENINFO ioctl with
transparency value.

Signed-off-by: Jingoo Han <jg1.han@samsung.com>