V4L/DVB: s5p-mfc: Change cacheable interface in decoder
This patch changes cacheable interface in decoder.
- 0 : nothing
- 1 : Cacheable for destination buffer
- 2 : Cacheable for source buffer

Before this patch, cacheable setting value was always overwritten.
To enable the cacheable for both source and destination, setting value
should be 3.
After this patch, cacheable setting value is calculated OR operation.
To enable the cacheable for both source and destination, separate
settings are needed as 1 after 2.

Change-Id: I53a1dc2a39791fe93e59d99f9207b654c4339054
Signed-off-by: Janghyuck Kim <janghyuck.kim@samsung.com>
Signed-off-by: Jinsung Yang <jsgood.yang@samsung.com>