iommu: exynos: replace spinlock with rcu lock in iovmm
exynos-iovmm protects the list of IOVMMs with spinlock and a IOVMM of
each device created with iovmm_setup(). The previous implementation
did not provides multiple setup a single IOVMM. This patch protects
the list of IOVMM from multiple setup from a device as well as race
condition from different devices without spin lock.

It is safe to use RCU lock instead of spinlock:
 - Call to IOVMM API functions are safe unless the state of a IOVMM
   of a device is changed while calling the functions. RCU lock does
   not change the state of protected data while a reader accesses
   the data.
 - An IOVMM does not need to be destroyed as soon as iovmm_cleanup()
   is called.

Change-Id: Ib3d29479bbf850edda3cd8142e01e79d863f1ca0
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>