media: gsc: add gscaler Rev1.0 variants
Gscaler varinats in GAIA Rev1.0 are different from Rev0.0.
So, the function to distinguish Rev1.0 from Rev0.0 is required.
And gscaler minimun size, cropped source size ,input buffer
count and so on have changed depending on gscaler version.
This patch added gscaler Rev1.0 variansts and a function to support
Rev1.0.

Change-Id: I101f1e448eaa4043690ef3428f16c070a3272188
Signed-off-by: Hyunwoong Kim <khw0178.kim@samsung.com>