media: vb2-ion: refactor vb2-ion and add support for internal buffers
Refactoring vb2-ion: for explicit and concise understande of vb2-ion.
- vb2_ion_init -> vb2_ion_create_context
- vb2_ion_cleanup -> vb2_ion_destroy_context
- vb2_ion_resume -> vb2_ion_attach_iommu
- vb2_ion_suspend -> vb2_ion_detach_iommu

Changed the meaning of vb2_plane_cookie() to return really a cookie.
Since vb2_plane_cookie() does not return a device address but a cookie
of the specified buffer, callers of it must care about the fact.
With the cookie returned by vb2_plane_cookie() you can:
- retrieves DMA address with vb2_ion_dma_address()
- retrieves physical memory information with vb2_ion_get_sg()

This patch also the drivers that need internal buffers such as MFC.
- vb2_ion_private_alloc() allocated a buffer and returns a cookie
- vb2_ion_private_free() accepts a cookie and frees the buffer
- vb2_ion_sync_for_cpu() and vb2_ion_sync_for_device() clean or
  invalidate cache lines of the specified area.
- vb2_ion_private_vaddr() returns the beginning virtual address of
  the allocated buffer.

Change-Id: If666aad455590839fb758aed27aa9cc8e4167829
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>