gpu: ion: exynos: enhance performance of user mapping by exynos-heap
This changes the mapping function of exynos-heap for userspace from
vm_insert_page() to remap_pfn_range(). Advantages of userspace mapping
with vm_insert_page() are not helpful for Exynos port of ION and
remap_pfn_range() is much faster than vm_insert_page() due to the
following reasons:
1. remap_pfn_range() maps large contiguous page at once while
   vm_insert_page() maps just a page at once.
2. remap_pfn_range() does not care about the states of pages it is
   mapping to user while vm_insert_page() checks and updates the state
   of the page it maps.
3. Compound pages must be split before giving them to vm_insert_page()
   while remap_pfn_range() does not care about it.

Splitting compound pages are not removed because UMP still needs it.

Change-Id: I60d95232abaa245e34bc5442629cced8ffd1d8cd
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>