iommu: exynos: print name of System MMU when fault occurs.
This patch adds a argument to system MMU's fault handler that
receives the name of the System MMU that generates fault.
Thus, the default fault handler prints the name of the System MMU
that made fault.

This also modifies G2D driver that is the only device driver
that implements fault handler.

Change-Id: I76472c715177765da273797fa6e9b9b7ef21a883
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>