iommu: exynos: remove lock for page table
This function removes spin_lock for page table from updating mapping
information in race condition since it is actually no effect if an IO
virtual memory manager is correct. If a faulty IO virtual memory
manager is removing some of the region that is being mapped, it makes
a serious problem of inconsistent mapping information remained in the
page table that is hard to debugging.

Removing spin-lock for page table also results in removal of callback
function of iommu_iova_to_phys() because it requires that the state of
the page table is consistent and that no update is performing.

Change-Id: Ib22b0e3064028df44a29bed21f3335114a283646
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>