media: gsc-out: remove unnecessary waiting function in FIMD ISR
GSC & FIMD stop sequence is like the below.
- Waiting for Vsync -> GSCALER stop -> FIMD stop.
But the clock of gscaler shouldn't be off until gscaler runs.
So, the function to wait until gscaler status would be idle is
added into FIMD ISR. This function takes about 20ms for waiting.
So, This patch moved this function from FIMD ISR to gscaler
stream off function.

Change-Id: I0f5089be4a84bd076174a02b9b0e43283faf3138
Signed-off-by: Hyunwoong Kim <khw0178.kim@samsung.com>