ARM: S5P: Add to change QoS during runtime.
This patch adds the implement to change QoS during runtime.
The platform data defines the change actions and passes it to BTS driver.
BTS driver implements a APIs, bts_set_priority() for it.
And then, The drivers that needs to change QoS during runtime
should call bts_set_priority() API.

Change-Id: I4d5d0773cb3248e82f9311436812e692e5e1dcf0
Signed-off-by: Boojin Kim <boojin.kim@samsung.com>