V4L/DVB: s5p-mfc: Modify DPB size calculation
This patch modifies DPB size calculating function.

Macroblock number of height should be multiple of 2 if stream is
interlaced mode. This alignment is conditionally applied in DPB size
calculating function in case of resolution is less than 2M pixel.
This condition checking is not necessary, so it is removed. And the
alignment is applied unconditionally.

Change-Id: Ie0b9319e27a152f2ba0e20be92cc45de0adffeb2
Signed-off-by: Janghyuck Kim <janghyuck.kim@samsung.com>