media: mipi-csis: applied optional define for sensor Jpeg
Applied optional define to choose sensor Jpeg in camera module S5K4ECGX
or HW jpeg in AP(Exynos) for Jpeg capture image.
If code of media bus is V4L2_MBUS_FMT_JPEG_1x8,
it can support both version,  Jpeg in camera module and Jpeg in
AP with defining JPEG in kernel menuconfig.

In the media device concept, formats and sizes of the IPs are matched in
the pipeline for external ISP.

Change-Id: Ib038df74bade742c7467413200103550aa464175
Signed-off-by: Tonghun Hwang <thun.hwang@samsung.com>