media: s5k4ecgx: Applied optional define for sensor Jpeg
S5K4ECGX supports Jpeg format and YUV format as output format
of the sensor module. the output connected FIMC-Lite DMA output
through the pipeline using media device concept.
In media device concept, sizes and formats of the IPs in the pipeline
must be matched. In the drivers, there are the functions finding
formats and sizes using the code for media bus format.

The optional define make user decide format which they need
for the sensor output.

Change-Id: I07661974caa8a834fd2e03ddd038a114e437d41c
Signed-off-by: Tonghun Hwang <thun.hwang@samsung.com>