ARM: EXYNOS: erased defining external interrupt for Ext.ISP
Erased the code defining external interrupt for
camera sensor module including ISP, M5MOLS.
Exynos5 does not need the external interrupt
because, in Exynos5 solution, the mainline v4l2 dirver for
camera is used.

This patch also make it avoid the conflict with USB 2.0
in host mode. The conflict caused no response to input
of the mouse device using USB 2.0.

Change-Id: Ie32b0d913c10a6dbfad3f2029a01b4fadd852ab6
Signed-off-by: Tonghun Hwang <thun.hwang@samsung.com>