media: s5k4ecgx: Added preview sizes
Added preview mode frame sizes for the format of YUV
as output of the sensor.
The YUV image format is used for capture image and
it would be changed its color format by Gscaler and
its data format by Jpeg.
There are also capture mode frame sizes for YUV format,
but capture mode frame image does not applied AE/AWB/AF,
when it have the sequence that sensor power shuts off and on
to change image size and format.

Change-Id: I8944c1c5e42897edf6cf2b640a7d3d7890633acc
Signed-off-by: Tonghun Hwang <thun.hwang@samsung.com>