gpu: vithar: Correctly update core state when suspending
The coarse-mode power policy didn't update the availability
bitmap when it turned the GPU off. So once the GPU was
turned on the PM core would think the shader cores were
powered when in reality they were not.

This patch causes the PM core to update the bitmap when the
coarse-mode policy turnes off the GPU to indicate the true state of
the hardware.

Change-Id: I818b13332fa5b56b941ae77f6ef2e7218d912a95
Signed-off-by: Kim Hyuncheol <host.kim@samsung.com>