gpu: vithar: return correct value from pm_callback_power_on
After a power cycle of the GPU the chicken bits will have become unset.
The callback to power up the GPU needs to notify the driver that
this has happened so that the necessary chicken bits can be set again.
In case of GPU cycled, return 1 and If GPU is powered on, return
from pm_callback_power_on.

Change-Id: I286342c99ac587718692bd1af411910a65120016
Signed-off-by: Kim Hyuncheol <host.kim@samsung.com>